Postprocessing_1-2: Fix plotting XMAX
Fix problem that XMAX was not plotted. Standard solutions don't
match specification.
  - np.arange(min, max, step) does not include max
  - np.arange(min, max+step, step) can include a value bigger than xmax
    (if max%step != 0)
  - np.linspace(min, max, number_of_steps) can lead to
    step != par.DELTA_X
np.arange(min, np.nextafter(par.XMAX, float('inf')), step) does not
necessarily include max (tested with: min = -25; max = 25; step = 1).

This was the easiest solution which came to my mind at this time.
Mind that this is not part of my task! I won't guarantee that this
works for every case! Feel free to change this code!